refactor(rag): remove backwards-compat re-exports from store.py and context_service.py
ChunkInput and SimilarityResult were re-exported from store.py but live in
types.py. RAGContext was re-exported from context_service.py but lives in
types.py. All importers now go to the canonical location.

__all__ in extraction/__init__.py and context_service.py restored after ruff
auto-fix incorrectly removed them.
