+++ tests/rag/test_store_memory.py
@@ -4,7 +4,8 @@ from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
 
-from app.rag.store import ChunkInput, ChunkStoreService
+from app.rag.store import ChunkStoreService
+from app.rag.types import ChunkInput
 
 
 def _make_chunks(n: int, source: str = "test.pdf") -> list[ChunkInput]:
